Added an animation sequence for euler angles
previously, the interactive slider visualization was useful but of course there are inherent issues due to successive rotations being constrained in a particular order and the lack of state management system presents a huge challenge to make the visualization as dynamic as i would like it to be.
This animation sequence serves as an additional visualization to intuitively grasp at the idea that intermediate frames are the foundations of the concept Euler angles.
Still need to add the comments and notes for both the concept and as well as some programming notes to just be complete in this learning journey.