Removed interactive visualization
Removed interactive visualization as it was getting complicated to solve it within jupyter notebook. Maybe the interactive element is better in a standalone gui and is not the best within the notebook.

Checked the final E-frame orientation and tested it against the results from 3D rotation converter, for now, i have printed the rotation matrix in a transposed way to match the output of the online converter but i have to check the proper output format for the final DCM. Find out this point and add it into the notebook as a key point.

Still have yet to add kinematic differential equation notes in euler angle note.

Start conceptualizing PRV visualizations.