Edited apply_rotations and Euler_rotation_animation
Previously, the target angles were input as 3 separate integers. Then these values are mapped into a dictionary.
However, for a proper euler angle sequence like '313' or '323', the mapping goes nonsensical as a dictionary requires a unique key identifier. This issue cannot be observed in a asymmteric sequence like '321' as there is no repetition of axis rotations and all the target angles are mapped to a unique key  identifier, and the identifiers are obviously the axis of rotation.

To overcome this issue, i changed the target angles input to a singular tuple, with 3 elements in the tuple, the first element of the tuple corresponds to the first rotation of the specified axis etc...
This effectively solved the problem. It was a small issue that had cascading consequences.

Learning point: Be aware of data types and their uses