Corrected Euler Rotation Animation
Suspected the rotations and DCMs were wrong. Initial suspicion was due to the wrong DCM that was being printed out in the Test Code Notebook.

The mistakes made were:
1) Wrong Rotation Matrix Definition (MAJOR)
The definition was taken with respect to wikipedia and ChatGPT. Both these sources gave a matrix that was wrong, despite the sources claim that these matrices are define for a right-handed frame.
When in doubt, re-visit first principles and derive the matrices by hand. Read the reference text as well, the conventions are very important.

2) Unnecessary transposition. There was no need for transposition after setting the conventions proper, now there is no transposition taking place when printing out the DCM or reading the DCM when adding to the frames.

Note: Forward Mapping DCMs, if defined to go from inertial frame to body frame, is in row major order. Once again, when in doubt, write out the matrices by hand to see if the elements of the matrix are in row or column major. They make a huge difference in the meaning.
Typically the transpose of the DCM will be read in column major as it is being reverse-mapped.