Updated Kinematics Notebook
Added a comment in the apply_rotations() function detailing the order of matrix multiplication.

Note: Matrix multiplication is non-commutative and you have to be very clear in the context of attitude dynamics, which vector/matrix you are transforming. Consider Matrix BN, if it is a DCM that transforms a matrix or a vector from one frame to another, then it is multiplied as such,
{v_body} = [BN] {v_inertial}
and the reason why it is as such, instead of post-multiplying v_inertial vectrix is due to the fact that for this notebook, the convention is row major for [BN]. So each row serves technically as the rotated frames' unit vectors!