Completed PRV Animation code in Test Code Notebook
Pretty much done with PRV animation.

Things i have identified:

1) DCM is literally the mother of all attitude representation.

So, maybe consider writing functions in a separate .py script to convert between, Euler angles and PRV to DCM, vice versa. Design such that once the DCM is obtained, the rest of the functions will take DCM as the input?
As long as there is a DCM, the attitude representation can be derived with a bit of math