Identified a major error in Euler angle code. Luckily this error isnt in the main visualization function
Identified a major error in the pre-refactored code. The output was not conforming to right hand positive rotations. So when the animation was played, a clockwise rotation about each axis was observed. However, the convention is counter-clockwise (right hand rule) that is positive angular displacement.
The function definitions were fine and all.
Thus the code that was developed before defining the functions was wrong. Furthermore, the logic was overly complicated and a simpler logic would have been enough to maintain the same output.
So, using the logic written in apply_rotations function, i rewrote the pre-refactored code which is meant to serve as prototype code that readers can use to understand the purpose of function definitions and refactoring; to create good readable code.