Created a section for Concept Check Questions that requires code to solve
Concept Checks covered in this notebook, respective to each section, is only solved for questions that require code to solve.
Readers should be able to solve other theoretical questions in the coursera course by means of reference to either this notebook, the textbook, or the slides and lectures provided in the course.