Began working on Quaternion Visualization Test code
After a brief hiatus, i have begun working on the visualization of quaternions.
The visualization will be very similar to PRVs.
For this visualization i hope to exploit chatgpt's o1-preview model. Of course it cannot be just cut and paste, the code will nee d to be well documented for readability and most importantly, understanding.