Added the polhode plotter function
Note that for large values of I, which can be representative of real life spacecrafts and or other rigid bodies, the Polhode Plotter somehow breaks and no curve is plotted out.

I suspect it to be due to sparse number of points in the grid that identifies the intersecting points. Increasing num points variable is impractical due to increased computation time, and not to mention the fact that it leads to Memory Error due to excessive RAM usage.
For now, keep it as it is, but investigate the use of numba or CuPy to accelerate the computation of intersecting points.