Create Test notebook for Polhode Plotter Function.ipynb
In this notebook, i have changed the portion of the code within the polhode_plotter function that creates the mesh grid for momentum sphere.
Instead i have opted to use my knowledge on the geometry, since the momentum sphere, is essentially a sphere, we can leverage on the use of spherical coordinates.
Then the resolution of our points will essentially boil down to the number of angular points across the azimuthal and elevation directions. Thus the computational effort reduces from a cubic complexity to a quadratic complexity, drastically reducing the computational and thus time expended to solve for the intersection of sphere and ellipsoid.

Furthermore, the previous method was essentially a brute force search to find grid points, and its a 3d grid. So even with parallelized vector computation using NumExpr Library, it took a significant amount of time. Now, having reduced the problem to a quadratic, or rather 2D grid problem across theta and phi, the time reduction coupled with the fact that the points are found with a single constraint, that is if the points are just outside of the surface of the ellipse with a tolerance close to 0, it will be the intersection point. So these points, appended to a list will form the curve of intersection. So even for large I values, where the sphere and ellipsoid scales, the intersection is still clearly seen as we are not focused on finding common points between 2 huge sets of 3 dimensional grids.