Update DCM_utils.py
Update Skew Symmetric matrix to handle 4 element vector and to output a 4x4 skew symmetric matrix.
This will be useful when dealing with Euler Parameters' Differential Equation to compute q_dot.
One of the forms is:
- q_dot = 0.5 * skew_symmetric(q) * omega
OR
- q_dot = 0.5 * skew_symmetric(omega) * q

where both omega and q is a 4 element vector.
Omega in this operation will have its first value as 0.