Update EOP Visualization.ipynb
Imported AttitudeKinematicsLib into notebook.
Before proceeding to full-blown animation, it would be wise to revisit the folder "1 - Kinematics: Describing the motions of spacecraft" and re-analyze how the frame visualizations were done previously.
Then maybe it might be worthwhile to create .py script for the plotly animation. Of course make it modular and re-usable.